requirements.txt
numpy
tifffile
openslide-python
